Move to types to specific files and code style changes.
