+++ src/NCalc/Domain/BinaryExpression.cs
@@ -18,27 +18,3 @@ public class BinaryExpression(
         visitor.Visit(this);
     }
 }
-
-public enum BinaryExpressionType
-{
-    And,
-    Or,
-    NotEqual,
-    LesserOrEqual,
-    GreaterOrEqual,
-    Lesser,
-    Greater,
-    Equal,
-    Minus,
-    Plus,
-    Modulo,
-    Div,
-    Times,
-    BitwiseOr,
-    BitwiseAnd,
-    BitwiseXOr,
-    LeftShift,
-    RightShift,
-    Unknown,
-    Exponentiation
-}
