+++ src/NCalc/Domain/BinaryExpressionType.cs
@@ -1,0 +1,25 @@
+namespace NCalc.Domain;
+
+public enum BinaryExpressionType
+{
+    And,
+    Or,
+    NotEqual,
+    LesserOrEqual,
+    GreaterOrEqual,
+    Lesser,
+    Greater,
+    Equal,
+    Minus,
+    Plus,
+    Modulo,
+    Div,
+    Times,
+    BitwiseOr,
+    BitwiseAnd,
+    BitwiseXOr,
+    LeftShift,
+    RightShift,
+    Unknown,
+    Exponentiation
+}
