+++ src/NCalc/Domain/UnaryExpressionType.cs
@@ -1,0 +1,9 @@
+namespace NCalc.Domain;
+
+public enum UnaryExpressionType
+{
+    Not,
+    Negate,
+    BitwiseNot,
+    Positive
+}
